requirements.txt
pyTelegramBotAPI
python-dotenv
requests
qrcode
Pillow 